+++ packages/adapters/nestjs/jest.config.ts
@@ -6,9 +6,9 @@ export default async (): Promise<Config.InitialOptions> => {
 
   return {
     ...baseConfig,
-    name: 'nestjs',
-    displayName: 'nestjs',
-    collectCoverageFrom: ['src/**/*.ts'],
-    coveragePathIgnorePatterns: ['index.ts'],
+    name: 'adapters.nestjs',
+    displayName: 'adapters.nestjs',
+    collectCoverageFrom: ['src/**/*.ts', '__test__/**/*.ts'],
+    coveragePathIgnorePatterns: ['index.ts', '__test__/integration.assets.ts'],
   };
 };