+++ packages/common/src/index.ts
@@ -1,7 +1,13 @@
 import { Type } from '@automock/types';
 
-export type ClassDependencies = Map<Type | string, Type>;
+export type PrimitiveValue = string | number | boolean | symbol | null;
+
+export type ClassDependencies = [Type | string, PrimitiveValue | Type][];
+
+export interface ClassDependenciesMap {
+  constructor: ClassDependencies;
+}
 
 export interface DependenciesReflector {
-  reflectDependencies(targetClass: Type): ClassDependencies;
+  reflectDependencies(targetClass: Type): ClassDependenciesMap;
 }