+++ packages/core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.2.0 (2023-06-10)
+
+### Features
+
+- **core:** support primitive value in testbuilder api ([#59](https://github.com/omermorad/automock/issues/59)) ([ef01f7c](https://github.com/omermorad/automock/commit/ef01f7ccc95867c66f992e78d7de90c353e53671))
+
 # 1.1.0 (2023-06-02)
 
 **Note:** Version bump only for package @automock/core