+++ packages/testbeds/jest/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/omermorad/automock/compare/@automock/jest@1.2.0...@automock/jest@1.2.1) (2023-06-24)
+
+**Note:** Version bump only for package @automock/jest
+
+
+
+
+
+# [1.2.0](https://github.com/omermorad/automock/compare/@automock/jest@1.0.1...@automock/jest@1.2.0) (2023-06-10)
+
+**Note:** Version bump only for package @automock/jest
+
 # [1.1.0](https://github.com/omermorad/automock/compare/@automock/jest@1.0.1...@automock/jest@1.1.0) (2023-06-02)
 
 ### Features